pyproject.toml
[tool.poetry]
name = "dialectical-framework"
version = "0.2.0"
description = "A dialectical framework for augmented intelligence. AI reasoning powered with dialectics supports humans in: system optimization (psychology, engineering, business, politics, etc.); dispute resolution (mediation, conflicts, negotiations, etc.); decision-making (dilemmas, challenging situations, win-win, etc.)."
authors = ["Evaldas Taroza <evaldas@dialexity.com>"]
readme = "README.md"

[[tool.poetry.packages]]
include = "dialectical_framework"
from = "src"



[tool.poetry.dependencies]
python = "^3.11"
mirascope = {extras = [
    "azure",
    "openai",
    "anthropic",
    "litellm",
    "bedrock",
    "langfuse"
], version = "^1.25.6"}
python-dotenv = "^1.0.1"
tabulate = "^0.9.0"
notebook = "^7.4.4"
dependency-injector = "^4.48.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
isort = "^6.0.1"
#flake8 = "^7.1.2"
pytest-asyncio = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"